requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-cloud-firestore==2.16.0
google-genai==1.10.0
google-cloud-pubsub==2.21.1
google-auth==2.30.0
google-cloud-secret-manager==2.20.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx>=0.28.1
aiofiles==23.2.1
